Reorganized the project. Removed desktop_side since this is no longer necessary because of direct interaction with ros master. Tested ultrasonic sensor publish.
