+++ apps/node2/node2.cpp
@@ -1,4 +1,5 @@
 #include "node2.h"
+
 #include "rcl.h"
 #include "Node.h"
 #include "Publisher.h"
