--- asw/apps/xmlrpc/xmlrpc.h
@@ -1,1 +1,0 @@
-void xmlrpc_task(void* p);
